[JDC-47-Rename-Calculator] Rename Calculator to main-calculator. Fixed #47
