+++ src/calculator/chains/qian-jin-zhui.ts
@@ -1,4 +1,4 @@
-import { SkillChainPayload } from '@calculator/calculator';
+import { SkillChainPayload } from '@calculator/main-calculator';
 import {
   increasedSansheng,
   insertPayloadSkills,
